replace zig tests with upstream testing framework
Libgit2 uses a handrolled testing framework called Clar. This patch compiles
and runs the core library tests. I've already caught a bug in the build script
from this so it's well worth it.

The current failure detection/reporting is pretty brittle, so I want to replace it
with a custom build step which will stream and parse TAP output from the runner.
This could hook into the progress and testing APIs for friendlier reporting as well.