update to zig 0.15.1
Fixes #9, supercedes #8

The test fixtures/resources are now copied to avoid modifying the global cache,
requiring a standalone build step to modify the file permissions.

Same goes for the `clar` source.
The previous ~hack~implementation worked since the libgit dependency path,
and by extension the `test/resources` subpath, was known at configure-time.
It was also an obvious misuse of the build API.

Now that it depends on the final fixtures path it's not possible to pass the
flag at configure-time. So instead a standalone build step replaces *reads* of the
`CLAR_FIXTURE_PATH` macro in the source with the final path. Two bugs, one stone.

Stdio polling was removed from `ClarTestStep` for simplicity.
It's probably more efficient too since it's not spinning on a syscall.

We'll have to wait until the MbedTLS dependency is updated to zig 0.15.1 before merging.